# contracts/openapi.yaml
openapi: 3.0.3
info:
  title: PLoT Engine API
  version: 1.0.0
  description: |
    PLoT Engine v1 - Deterministic, auditable probabilistic models with trust signals.
    
    **Trust Signals**: All /v1 responses include Model Card, Confidence badge, and Explain-Δ.
    
    **Demo Mode**: Add `?demo=1` or header `X-Demo: 1` for deterministic sample payloads.
    When present, demo mode short-circuits request validation and returns a canonical demo payload.
    
    **Feature Flags**:
    - FEATURE_STREAM=1 enables real SSE /stream route (test-only otherwise)
    - AUTH_ENABLED=1 requires Authorization: Bearer <token> on certain routes
    
    **Legacy Routes**: /draft-flows, /critique, /improve (v0) remain for backwards compatibility.
  contact:
    name: PLoT Engine
    url: https://github.com/Talchain/plot-lite-service
    email: devnull@talchain.ai
servers:
  - url: http://127.0.0.1:4311
    description: Local development
  - url: https://api.olumi.ai
    description: Production (future)
...
paths:
...
  /v1/run:
    head:
      summary: Probe endpoint availability
      operationId: headRunV1
      tags: [v1, meta]
      description: |
        Returns 204 with no body. Used by UI clients to verify endpoint availability.
      responses:
        '204':
          description: No Content - endpoint is available
    post:
      summary: Execute probabilistic model
      operationId: runV1
      tags: [v1, inference]
      description: |
        Run inference on a causal graph with full trust signals.
        Demo mode: `?demo=1` short-circuits validation and returns a canonical deterministic sample.
      parameters:
        - name: demo
          in: query
          required: false
          schema: { type: integer, enum: [0, 1] }
        - name: X-Demo
          in: header
          required: false
          schema: { type: string, enum: ['0', '1'] }
        - name: Idempotency-Key
          in: header
          required: false
          description: Optional key to replay the exact prior response within a TTL (scoped per principal)
          schema:
            type: string
            maxLength: 200
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/runRequest'
            examples:
              canonical_targets:
                summary: Canonical targets field
                value:
                  graph:
                    nodes:
                      - id: "A"
                        label: "Driver"
                        belief: 0.6
                      - id: "B"
                        label: "Outcome"
                    edges:
                      - id: "e1"
                        from: "A"
                        to: "B"
                        weight: 0.7
                        rationale: "A pushes B"
                  targets: ["B"]
                  seed: 4242
              legacy_query_targets:
                summary: Legacy query.targets (deprecated)
                value:
                  graph:
                    nodes:
                      - id: "A"
                        label: "Driver"
                        belief: 0.6
                      - id: "B"
                        label: "Outcome"
                    edges:
                      - id: "e1"
                        from: "A"
                        to: "B"
                        weight: 0.7
                        rationale: "A pushes B"
                  query:
                    targets: ["B"]
                  seed: 4242
      responses:
        '200':
          description: OK
          headers:
            Idempotent-Replayed:
              description: '1 if response is a replay from cache; 0 if computed fresh (when Idempotency-Key provided)'
              schema: { type: string, enum: ['0','1'] }
            X-RateLimit-Limit:
              description: Maximum requests per minute
              schema: { type: string }
            X-RateLimit-Remaining:
              description: Remaining requests in current window
              schema: { type: string }
            X-RateLimit-Reset:
              description: Unix timestamp when rate limit resets
              schema: { type: string }
            X-SCM-Lite:
              description: SCM-Lite enablement flag (0 or 1)
              schema: { type: string, enum: ['0','1'] }
            X-Olumi-Backend:
              description: Active backend mode (fallback or scm_lite)
              schema: { type: string, enum: ['fallback','scm_lite'] }
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/runResponse'
              examples:
                success:
                  value:
                    schema: run.v1
                    result:
                      response_hash: "a1b2c3d4e5f6..."
                      summary:
                        p10: 0.15
                        p50: 0.42
                        p90: 0.78
                      confidence: 0.85
                    meta:
                      seed: 4242
                      k_samples: 256
        '413':
          description: Payload too large
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/errorV1'
              examples:
                too_large:
                  value:
                    schema: error.v1
                    code: BAD_INPUT
                    message: Request entity too large
        '429':
          description: Too many requests (per-minute limit)
          headers:
            Retry-After:
              description: Seconds to wait before retrying
              schema: { type: integer, minimum: 0 }
            X-RateLimit-Reset:
              description: Unix epoch seconds when the window resets
              schema: { type: string }
            X-RateLimit-Reason:
              description: Reason for rate limit decision
              schema: { type: string, enum: [per_ip] }
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/errorV1'
              examples:
                rate_limited:
                  value:
                    schema: error.v1
                    code: RATE_LIMITED
                    message: Too many requests
        '400':
          description: Bad input
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/errorV1'
              examples:
                bad_input:
                  value:
                    schema: error.v1
                    code: BAD_INPUT
                    message: 'Missing required field: graph'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/errorV1'
              examples:
                unauthorized:
                  value:
                    schema: error.v1
                    code: UNAUTHORIZED
                    message: Missing bearer token
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/errorV1'
              examples:
                forbidden:
                  value:
                    schema: error.v1
                    code: FORBIDDEN
                    message: Invalid token
...
  /v1/sensitivity:
    post:
      summary: Sensitivity & Robustness Analysis
      description: |
        One-at-a-time (OAT) sensitivity analysis with tornado charts.
        Perturbs each target node by ±delta and measures effect on outcome.
      operationId: sensitivity
      tags: [v1, analysis]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [graph]
              properties:
                graph:
                  $ref: '#/components/schemas/graph'
                seed:
                  type: integer
                  example: 4242
                method:
                  type: string
                  enum: [oat]
                  default: oat
                  description: Sensitivity method (one-at-a-time)
                delta:
                  type: number
                  minimum: 0
                  maximum: 1
                  exclusiveMinimum: true
                  default: 0.1
                  description: Proportional perturbation for OAT
                targets:
                  type: array
                  items:
                    type: string
                  description: Optional subset of node_ids to analyze (default all)
            example:
              graph:
                nodes:
                  - {id: "A", label: "Node A", value: 0.5}
                  - {id: "B", label: "Node B", value: 0.3}
                edges: []
              seed: 4242
              method: "oat"
              delta: 0.1
              targets: ["A", "B"]
      responses:
        '200':
          description: Sensitivity analysis results
          content:
            application/json:
              schema:
                type: object
                properties:
                  schema:
                    type: string
                    example: "sensitivity.v1"
                  summary:
                    type: object
                    properties:
                      baseline:
                        type: object
                        properties:
                          p50:
                            type: number
                  drivers:
                    type: array
                    items:
                      type: object
                      properties:
                        node_id:
                          type: string
                        tornado:
                          type: object
                          properties:
                            low:
                              type: number
                              description: Effect when node decreased by delta
                            high:
                              type: number
                              description: Effect when node increased by delta
                        rank:
                          type: integer
                          description: Rank by total swing (1 = highest)
                  meta:
                    type: object
                    properties:
                      seed:
                        type: integer
                      method:
                        type: string
                      evaluations:
                        type: integer
                        description: Number of model evaluations (1 + 2*N)
              example:
                schema: "sensitivity.v1"
                summary:
                  baseline: {p50: 0.624}
                drivers:
                  - {node_id: "A", tornado: {low: -0.06, high: 0.08}, rank: 1}
                  - {node_id: "B", tornado: {low: -0.03, high: 0.04}, rank: 2}
                meta:
                  seed: 4242
                  method: "oat"
                  evaluations: 5
        '400':
          description: Bad input
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: object
                    properties:
                      type:
                        type: string
                      message:
                        type: string
              examples:
                invalid_delta:
                  value:
                    schema: "error.v1"
                    code: "BAD_INPUT"
                    message: "delta must be in (0, 1]"
                invalid_targets:
                  value:
                    schema: "error.v1"
                    code: "BAD_INPUT"
                    message: "Invalid target node_ids: Z"
...
components:
  schemas:
...
    errorV1:
      type: object
      required: [schema, code, message]
      properties:
        schema:
          type: string
          enum: [error.v1]
        code: { type: string, enum: [BAD_INPUT, UNAUTHORIZED, FORBIDDEN, RATE_LIMITED, INTERNAL] }
        message: { type: string }
        path:
          type: array
          items: { type: string }
        details:
          type: object
          additionalProperties: true
...
    runResponse:
      type: object
      required: [schema, graph, results, model_card, confidence]
      properties:
        schema: { type: string, enum: [run.v1] }
        graph: { $ref: '#/components/schemas/graph' }
        results: { type: object }
        model_card: { $ref: '#/components/schemas/modelCard' }
        confidence: { $ref: '#/components/schemas/confidenceBadge' }
        linearity_warning: { type: object }
        threshold_crossings:
          type: array
          items: { type: object }
        fork_suggestions: { type: object }
        critique:
          type: array
          items: { $ref: '#/components/schemas/critiqueItem' }
        explain_delta: { $ref: '#/components/schemas/explainDelta' }
        identifiability: { type: string }
        trace_id:
          type: string
          description: 'Optional client-provided trace identifier echoed from X-Trace-Id when TRACE_ID_PASSTHROUGH=1; may also be present when TRACE_MIN=1'
        result:
          type: object
          description: 'P0: Convenience fields for UI integration'
          properties:
            response_hash:
              type: string
              description: 'SHA-256 hash of canonical inputs for caching'
            summary:
              type: object
              description: 'Direct percentile access'
              properties:
                p10: { type: number }
                p50: { type: number }
                p90: { type: number }
...